+++ rsa.h
@@ -1,0 +1,45 @@
+/**********************************************************************
+ *                                                                    *
+ * Created by Adam Brockett                                           *
+ *                                                                    *
+ * Copyright (c) 2010                                                 *
+ *                                                                    *
+ * Redistribution and use in source and binary forms, with or without *
+ * modification is allowed.                                           *
+ *                                                                    *
+ * But if you let me know you're using my code, that would be freaking*
+ * sweet.                                                             *
+ *                                                                    *
+ **********************************************************************/
+
+#ifndef RSA_H
+#define RSA_H
+#include <gmp.h>
+
+#define MODULUS_SIZE 1024                   /* This is the number of bits we want in the modulus */
+#define BLOCK_SIZE (MODULUS_SIZE/8)         /* This is the size of a block that gets en/decrypted at once */
+#define BUFFER_SIZE ((MODULUS_SIZE/8) / 2)  /* This is the number of bytes in n and p */
+
+typedef struct {
+    mpz_t n; /* Modulus */
+    mpz_t e; /* Public Exponent */
+} public_key;
+
+typedef struct {
+    mpz_t n; /* Modulus */
+    mpz_t e; /* Public Exponent */
+    mpz_t d; /* Private Exponent */
+    mpz_t p; /* Starting prime p */
+    mpz_t q; /* Starting prime q */
+} private_key;
+
+/* NOTE: Assumes mpz_t's are initted in ku and kp */
+void generate_keys(private_key* ku, public_key* kp);
+
+void block_encrypt(mpz_t C, mpz_t M, public_key kp);
+void block_decrypt(mpz_t M, mpz_t C, private_key ku);
+
+int encrypt(char cipher[], char message[], int length, public_key kp);
+int decrypt(char* message, char* cipher, int length, private_key ku);
+
+#endif
